Adopt, and adapt for, CASE 1.0.0
One potential bug has been flagged with this shape, implemented in UCO
Issue 406:
`uco-owl:ObjectProperty-shacl-constraints-shape`

The `sh:PropertyShape` raising the bug has been given an IRI in order to
link a deactivation rationale.

A new shapes file `debug.ttl` has been added to disable that shape until
a test is written to confirm the CASE-Corpora shape is correct.

`Facet`s that were blank nodes have been given IRIs, per the
implementation of UCO Issue 430.  New `sh:Info`-severity violations are
reported for some URLs treated in the "URL as an `rdfs:Resource` manner,
which will not be given UUID endings.  `case_validate` is called with
`--allow-warnings`, but is intended to be called with `--alow-infos`;
that will have to wait for `case-utils` Issue 70 to resolve.

Imports of CASE and UCO ontologies now use their `owl:versionIRI`s,
implemented in UCO Issue 437.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/casework/CASE-Utilities-Python/issues/70
* https://github.com/ucoProject/UCO/issues/406
* https://github.com/ucoProject/UCO/issues/430
* https://github.com/ucoProject/UCO/issues/437

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>