Migrate OWL transitive closure illustration script
As part of testing UCO Issue 491, I had developed a script to visualize
the transitive closure `owl:imports`.  The script resided in the
CASE-Examples-QC repository for reference, but was not operationalized.
Part of the documentation for CASE-Corpora is set to include a
visualization of the transitive closure of its various SHACL shape
graphs and upstream ontologies, so this patch migrates the script.

This file was migrated from the CASE-Examples-QC repository, commit:
de5d00d6933ed92309323a04f0d6b8f432715a63

References:
* https://github.com/ajnelson-nist/CASE-Examples-QC/
* https://github.com/ucoProject/UCO/issues/491

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>