Insert Observations and "witnessing" objects for consistent distribution files
This patch establishes a graph pattern, using a query in a unit test,
for setting ordering of file hash values based on files extracted from
the distribution tarball.  The pattern is initially applied to files
where hashes were consistent with the image documentation PDF.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>