Track and add normalization steps for SSN and SOSA ontologies
This patch starts the adoption process of integrating the Semantic
Sensor Network (SSN) ontology into CASE-Corpora's ontology and shapes.
The initial application of SSN is in representing contents of a file
being observed to change.

Due to an issue with a definition conflict between FOAF, SSN, and SOSA,
this initial patch sets up the generation record, but does not
integrate SOSA and SSN into the build process yet.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/w3c/sdw/issues/1395

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>