Define file hash quality to represent mutating file contents
This patch adopts SOSA Observations and Actuations as mechanisms for
representing state-at-time and modifications of file contents changing.
This is in support of review of a set of records about a file, with
different records indicating different file contents from hashes
changing.

This patch also defines a concept for future consideration of the UCO
Ontology Committee, a Quality, aligned with the SSN class Property.
Other ontologies are known to include Qualities as well (e.g. BFO and
UFO, with their respective classes linked using rdfs:seeAlso), and they
are planned to be tested using the CDO "Profiles" strategy (CDOTSC-29).

To maintain consistency with UCO's current usage of `ContentDataFacet`,
shapes are added to resolve conflicts between (SOSA+UCO) Observations
and hash values recorded in `Facet`s.

The first quality defined, `FileHashQuality`, intentionally chooses an
association of `observable:File` rather than `observable:ContentData`.
There is no subclass relationship in UCO between `observable:File` and
`observable:ContentData`, so it is not clear if a `ContentData` could
also appropriately consider its content-hashes to be modifiable (in
SOSA, "Actuatable") qualities.

This patch only defines the new classes, properties, shapes, and
demonstrative exemplars.  Usage in datasets is left for future patches.

Shapes for SOSA import shapes for FOAF due to:
1. a FOAF property re-definition included in SOSA that triggered an OWL
   conflict, and
2. usage of Dublin Core Terms, whose shapes are imported by the FOAF
   shapes.

A follow-on patch will regenerate Make-managed files.

References:
* [CDOTSC-29] UCO and/or CASE should create exploratory, non-binding
  "Profile" ontologies to test external ontology alignments

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>