Add Dublin Core shapes review to cover SKOS term usage
This patch adds SHACL shapes to review other properties used in one of
the imported upstream RDF schemata.

This patch also adjusts the response pattern to Dublin Core properties
being used in upstream ontologies in manners not conformant with the
suggestions written, but not encoded as RDFS, in Dublin Core.  Shapes
are now deactivated in the CASE-Corpora local shapes file.

No effects were observed on Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>