Add steps to download and normalize PROV-O
It has been an oversight that prov-o was not included in the transitive
closure construction, given its usage with `case_prov`.  The oversight
became a blocking issue in implementation of UCO Issue 449, which has a
solution in draft that requires the transitive closure be present, and
DCAT imports PROV-O as a dependency.

A follow-on patch will generate PROV-O per this recipe.

References:
* https://github.com/ucoProject/UCO/issues/449

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>