Generate supplemental Digital Corpora graph files
Some details entered manually for the Downloadable Files are now
delegated to being generated by Digital Corpora.

One side-effect is the consistency scheme for IRIs is starting to
migrate to spelling `owl:NamedIndividual` local names starting with the
spelling, including as capitalized, of their class's local name.
The brief rationale for this decision is that it saves on the effort of
devising for the eventual general CASE Python API a lowercasing,
word-segmenting procedure for `owl:NamedIndividual` IRI names that could
handle nontrivial cases, e.g. `uco-observable:ContentData` (going to
`kb:content-data-$uuid`), `uco-observable:WiFiAddress` (going to
`kb:wifi-address-$uuid`).

A follow-on patch will regenerate Make-managed files.  Note that even
though some `Hash` objects are deleted in this aptch, they are now
generated in the new supplemental file.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>