Add ProvenanceRecord to satisfy local shape
A local-to-CASE-Corpora shape requires `InvestigativeAction`s to have
`ProvenanceRecord`s as outputs.  (See `/shapes/local.ttl`,
`sh-case-corpora-local:InvestigativeAction-shape`.)
That shape used a SHACL mechanism that was not active in this
repository's testing until pySHACL Issue 213 was closed.

This patch adds a `ProvenanceRecord` as output to the last non-subaction
that had the phone as input.  Review of the results of this SPARQL query
indicate the device is not used in further actions, so this
`ProvenanceRecord` has no further impact on the graph today.

```sparql
SELECT ?nAction ?lDescription
WHERE {
  ?nAction
    uco-action:object kb:device-ea732801-7d0e-46ac-a028-69b782c97a46 ;
    .
  OPTIONAL {
    ?nAction
      uco-core:description ?lDescription ;
      .
  }
}
ORDER BY ?nAction
```

There is some open question on how to tie the subactions' outputs to the
parent action's `ProvenanceRecord`; that thread is on CASE Issue 136.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/RDFLib/pySHACL/issues/213
* https://github.com/casework/CASE/issues/136

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>