Define DownloadableObjects hierarchy and DownloadableRelation
This patch re-introduces `DownloadableFile`, as well as some related
classes, to describe a concept IRI that behaves as a file-yielding URL.

This patch removes a qualified SHACL constraint that had been written on
a conflated-classes guess that instigated UCO Issue 534.  After
discussion on Issue 534, this shape should have been removed, but it
wasn't until pySHACL Issue 213 was addressed that the shape started
triggering and raising awareness it was still around.

Hand-maintained data, as well as the Digital Corpora rendering script,
have been updated to accommodate the new shapes around
`DownloadableObject` and `DownloadableRelation`.  A later patch will
need to add the `DownloadableRelation`s to the chain of custody.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/RDFLib/pySHACL/issues/213
* https://github.com/ucoProject/UCO/issues/534

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>